+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+* FIX: MinimalJsonEncoder encoded \n,\t,\r,etc. (control characters that has a two-character sequence) fully escaped (\u000#)
+* FIX: MaximalJsonEncoder encoded \n incorrectly (\u000: instead of \u000a)
+
 ## 0.0.1
 * First release.